refactor: delete confirmation を global にする
