+++ workspaces/client/app/components/feature/delete-confirmation.tsx
